Renamed VERSION.py file to version.py.
